refactor(redteam): extract HTTP/error helpers to _errors.py
