+++ packages/evaluatorq-py/src/evaluatorq/redteam/backends/orq.py
@@ -43,7 +43,7 @@ def _get_orq_server_url() -> str:
     return url.rstrip('/').removesuffix('/v2/router')
 
 
-from evaluatorq.redteam.backends.base import extract_provider_error_code, extract_status_code
+from evaluatorq.redteam.backends._errors import extract_provider_error_code, extract_status_code
 from evaluatorq.redteam.contracts import (
     PIPELINE_CONFIG,
     AgentContext,
